requirements.txt
flask == 1.1.1
tqdm == 4.42.0
torch == 1.4.0
torchvision == 0.5.0
matplotlib == 3.1.2
gunicorn==20.0.4
numpy==1.18.0
Pillow==6.2.0
requests==2.22.0